[refactor/72-task] req -> request, res -> response
[refactor/72-task] req -> request, res -> response
