+++ ideac-admin/src/main/java/depth/main/ideac/domain/auth/presentation/AuthController.java
@@ -3,9 +3,9 @@ package depth.main.ideac.domain.auth.presentation;
 
 import depth.main.ideac.domain.admin.application.AdminService;
 import depth.main.ideac.domain.auth.application.AuthService;
-import depth.main.ideac.domain.auth.dto.res.AuthRes;
-import depth.main.ideac.domain.auth.dto.req.FindIdReq;
-import depth.main.ideac.domain.auth.dto.req.SignInReq;
+import depth.main.ideac.domain.auth.dto.response.AuthRes;
+import depth.main.ideac.domain.auth.dto.request.FindIdReq;
+import depth.main.ideac.domain.auth.dto.request.SignInReq;
 import depth.main.ideac.domain.user.dto.PasswordReq;
 import depth.main.ideac.global.payload.ErrorResponse;
 import depth.main.ideac.global.payload.Message;
