+++ ideac-core/src/main/java/depth/main/ideac/domain/auth/dto/request/FindIdReq.java
@@ -1,4 +1,4 @@
-package depth.main.ideac.domain.auth.dto.req;
+package depth.main.ideac.domain.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
