+++ ideac-core/src/main/java/depth/main/ideac/domain/auth/dto/request/RefreshTokenReq.java
@@ -1,10 +1,9 @@
-package depth.main.ideac.domain.auth.dto.req;
+package depth.main.ideac.domain.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
 
 @Getter
