+++ ideac-core/src/main/java/depth/main/ideac/domain/auth/dto/request/SignUpReq.java
@@ -1,11 +1,9 @@
-package depth.main.ideac.domain.auth.dto.req;
+package depth.main.ideac.domain.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
-import lombok.Data;
 import lombok.Getter;
 
 @Getter
