+++ ideac-core/src/main/java/depth/main/ideac/domain/auth/dto/response/AuthRes.java
@@ -1,8 +1,7 @@
-package depth.main.ideac.domain.auth.dto.res;
+package depth.main.ideac.domain.auth.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
 
 @Getter
