+++ ideac-user/src/main/java/depth/main/ideac/domain/user/presentation/UserController.java
@@ -1,8 +1,7 @@
 package depth.main.ideac.domain.user.presentation;
 
-import depth.main.ideac.domain.auth.dto.req.SignUpReq;
+import depth.main.ideac.domain.auth.dto.request.SignUpReq;
 import depth.main.ideac.domain.user.application.UserService;
-import depth.main.ideac.domain.user.dto.PasswordReq;
 import depth.main.ideac.global.config.security.token.CurrentUser;
 import depth.main.ideac.global.config.security.token.UserPrincipal;
 import depth.main.ideac.global.payload.ErrorResponse;
